+++ ui/src/pages/challenge-data/ChallengeSet1Data.js
@@ -1,32 +1,35 @@
-// import * as UI from '@material-ui/icons';
+// import * as UI from "@material-ui/icons";
 // import React, { useState, useEffect } from "react";
-// import ReactMarkdown from 'react-markdown';
-// import gfm from 'remark-gfm';
-// import marked from 'marked';
-// import fs from 'fs';
+// import ReactMarkdown from "react-markdown";
+// import gfm from "remark-gfm";
+// import marked from "marked";
+// import fs from "fs";
 
 //**********************************************cipher challenge set*****************************************
-import morsecode from "../challenges/ciphers/challenges/morse-code/docs/Morse_code.md";
-import crypto from "../challenges/ciphers/challenges/letter-to-number/docs/1-Cryptography.md";
-import term from "../challenges/ciphers/challenges/caesar-cipher/docs/Terminal_Use.md";
-export const Challengedata = [
+import ceasarcipher from "../../challenges/ciphers/challenges/caesar-cipher/docs/Caesar-cipher.md";
+import ltn from "../../challenges/ciphers/challenges/caesar-cipher/docs/1-Cryptography.md";
+import term from "../../challenges/ciphers/challenges/caesar-cipher/docs/Terminal_Use.md";
+
+export const ChallengeSet1Data = [
   {
     label: "Basic Terminal Commands",
     cName: "tab",
     itembox1: term,
   },
+
   {
     label: "Challenge Description",
     cName: "tab",
-    itembox1: morsecode,
+    itembox1: ceasarcipher,
     // <p style={{marginTop:'10px', marginLeft:'10px'}}  >
     //     {/* <ReactMarkdown remarkPlugins={[gfm]} children={markdown1} /> */}
     // </p>,
   },
+
   {
     label: "Cryptography",
     cName: "tab",
-    itembox1: crypto,
+    itembox1: ltn,
     // <p style={{marginTop:'10px', marginLeft:'10px'}}  >
     //     {/* <ReactMarkdown remarkPlugins={[gfm]} children={markdown1} /> */}
     // </p>,
