+++ ui/src/pages/challenge-data/ComprehensiveChallengeData.js
@@ -6,11 +6,11 @@
 // import fs from 'fs';
 
 //**********************************************cipher challenge set*****************************************
-import comprehensive from "../challenges/ciphers/challenges/comprehensive-challenge/docs/Mixed-Challenge-1.md";
-import crypto from "../challenges/ciphers/challenges/comprehensive-challenge/docs/1-Cryptography.md";
-import morse from "../challenges/ciphers/challenges/morse-code/docs/Morse_code.md";
-import alphaTable from "../challenges/ciphers/challenges/letter-to-number/docs/Letter-to-Number.md";
-import term from "../challenges/ciphers/challenges/caesar-cipher/docs/Terminal_Use.md";
+import comprehensive from "../../challenges/ciphers/challenges/comprehensive-challenge/docs/Mixed-Challenge-1.md";
+import crypto from "../../challenges/ciphers/challenges/comprehensive-challenge/docs/1-Cryptography.md";
+import morse from "../../challenges/ciphers/challenges/morse-code/docs/Morse_code.md";
+import alphaTable from "../../challenges/ciphers/challenges/letter-to-number/docs/Letter-to-Number.md";
+import term from "../../challenges/ciphers/challenges/caesar-cipher/docs/Terminal_Use.md";
 
 export const Challengedata = [
   {
