+++ ui/src/pages/challenge-data/MorseCodeData.js
@@ -6,22 +6,27 @@
 // import fs from 'fs';
 
 //**********************************************cipher challenge set*****************************************
-import dirtraverse from "../challenges/web-based/challenges/directory-traversal/docs/DirTrav.md";
-import dirtraverseDescription from "../challenges/web-based/challenges/directory-traversal/docs/DirTravDescription.md";
-
+import morsecode from "../../challenges/ciphers/challenges/morse-code/docs/Morse_code.md";
+import crypto from "../../challenges/ciphers/challenges/letter-to-number/docs/1-Cryptography.md";
+import term from "../../challenges/ciphers/challenges/caesar-cipher/docs/Terminal_Use.md";
 export const Challengedata = [
+  {
+    label: "Basic Terminal Commands",
+    cName: "tab",
+    itembox1: term,
+  },
   {
     label: "Challenge Description",
     cName: "tab",
-    itembox1: dirtraverseDescription,
+    itembox1: morsecode,
     // <p style={{marginTop:'10px', marginLeft:'10px'}}  >
     //     {/* <ReactMarkdown remarkPlugins={[gfm]} children={markdown1} /> */}
     // </p>,
   },
   {
-    label: "Directory Traversal",
+    label: "Cryptography",
     cName: "tab",
-    itembox1: dirtraverse,
+    itembox1: crypto,
     // <p style={{marginTop:'10px', marginLeft:'10px'}}  >
     //     {/* <ReactMarkdown remarkPlugins={[gfm]} children={markdown1} /> */}
     // </p>,
