+++ ui/src/pages/challenge-pages/CaeserCipher.js
@@ -1,19 +1,20 @@
 import React, { useState, useEffect } from "react";
-import "./ChallengeSet1.css";
+import "../../style/ChallengeSet1.css";
 import * as core from "@material-ui/core";
 //FIX THIS DATA FOR EACH FILE
 //CHALLENGE 1 = CAESAR CHALLENGE
 /*----------------------------------------------------------------------------- */
-import { Challengedata } from "./ComprehensiveChallengeData";
+import { ChallengeSet1Data } from "../challenge-data/ChallengeSet1Data";
 /*----------------------------------------------------------------------------- */
-// import {useLocation, Link, useParams} from 'react-router-dom';
+
 import PropTypes from "prop-types";
-import Terminal from "../components/Terminal";
+import Terminal from "../../components/Terminal";
+
 import gfm from "remark-gfm";
 import ReactMarkdown from "react-markdown";
-import useFetchAuth from "../useFetchAuth";
-import GenericErrorPage from "./error-pages/genericErrorPage";
-import Spinner from "../components/Spinner";
+import useFetchAuth from "../../useFetchAuth";
+import Spinner from "../../components/Spinner";
+import GenericErrorPage from "../error-pages/genericErrorPage";
 
 function a11yProps(index) {
   return {
@@ -46,26 +47,28 @@ TabPanel.propTypes = {
   value: PropTypes.any.isRequired,
 };
 
-export default function ComprehensiveChallenge() {
+export default function CaeserCipher() {
   const [value, setValue] = React.useState(0);
   const handleChange = (events, newValue) => {
     setValue(newValue);
-    setPath(Challengedata[newValue].itembox1);
+    setPath(ChallengeSet1Data[newValue].itembox1);
   };
 
+  // API POST REQUEST TO THE SERVER
   const { data, error, loading } = useFetchAuth(
     "api/user/environments",
     "POST",
-    { challengeSetSlug: "ciphers", challengeSlug: "comprehensive-challenge" }
+    { challengeSetSlug: "ciphers", challengeSlug: "letter-to-number" }
   );
-  const [txt, setMarkdown] = useState("");
-  const [path, setPath] = useState("");
-  const [first, setFirst] = useState(0);
+
+
 
   //LEARN THIS FOR FUTURE WORK:
   //USESTATE HOOK WILL TRIGGER WHEN THE ASSIGNED FUNCTION IS CALL (EXAMPLE BELOW IS SETMARKDOWN/ SETPATH)
   //USEEFFECT HOOK WILL TRIGGER THE WHEN THE VARIABLE AT THE [] AT BOTTOM IS UPDATE. IN THIS CASE IS "PATH VARIABLE"
-
+  const [txt, setMarkdown] = useState("");
+  const [path, setPath] = useState("");
+  const [first, setfirst] = useState(0);
   useEffect(() => {
     fetch(path)
       .then((res) => res.text())
@@ -73,12 +76,12 @@ export default function ComprehensiveChallenge() {
         setMarkdown(text);
       });
   }, [path]);
-  //this is to set first frim rendering for mardown. Without this, it will redner html on first run
   if (loading) return <Spinner />;
   if (error) return <GenericErrorPage />;
+  //this is to set first frim rendering for mardown. Without this, it will redner html on first run
   if (!first) {
-    setPath(Challengedata[0].itembox1);
-    setFirst(first + 1);
+    setPath(ChallengeSet1Data[0].itembox1);
+    setfirst(first + 1);
   }
   return (
     <div style={{ display: "flex", flexDirection: "row", position: "fixed" }}>
@@ -93,9 +96,8 @@ export default function ComprehensiveChallenge() {
             scrollButtons="auto"
             aria-label="simple auto tabs example"
           >
-            {Challengedata.map((item, index) => {
+            {ChallengeSet1Data.map((item, index) => {
               return (
-                // <core.Tab key={item.index} label={item.label} {...a11yProps(item.index)} component={Link} to={`${item.topic}/${item.label}`} />
                 <core.Tab
                   key={index}
                   label={item.label}
@@ -106,15 +108,12 @@ export default function ComprehensiveChallenge() {
           </core.Tabs>
         </core.AppBar>
 
-        {/* <ReactMarkdown source={Challengedata[value].itembox1}/> */}
-
         <TabPanel
           className="box1"
           value={value}
           index={value}
           style={{ overflowY: "scroll", marginTop: "5px", marginLeft: "5px" }}
         >
-          {/* {Challengedata[value].itembox1}  */}
           <ReactMarkdown
             remarkPlugins={[gfm]}
             children={txt}
