+++ ui/src/pages/challenge-pages/ComprehensiveChallenge.js
@@ -1,20 +1,19 @@
 import React, { useState, useEffect } from "react";
-import "./ChallengeSet1.css";
+import "../../style/ChallengeSet1.css";
 import * as core from "@material-ui/core";
 //FIX THIS DATA FOR EACH FILE
 //CHALLENGE 1 = CAESAR CHALLENGE
 /*----------------------------------------------------------------------------- */
-import { ChallengeSet1Data } from "./ChallengeSet1Data";
+import { Challengedata } from "../challenge-data/ComprehensiveChallengeData";
 /*----------------------------------------------------------------------------- */
-
+// import {useLocation, Link, useParams} from 'react-router-dom';
 import PropTypes from "prop-types";
-import Terminal from "../components/Terminal";
-
+import Terminal from "../../components/Terminal";
 import gfm from "remark-gfm";
 import ReactMarkdown from "react-markdown";
-import useFetchAuth from "../useFetchAuth";
-import Spinner from "../components/Spinner";
-import GenericErrorPage from "./error-pages/genericErrorPage";
+import useFetchAuth from "../../useFetchAuth";
+import GenericErrorPage from "../error-pages/genericErrorPage";
+import Spinner from "../../components/Spinner";
 
 function a11yProps(index) {
   return {
@@ -47,28 +46,26 @@ TabPanel.propTypes = {
   value: PropTypes.any.isRequired,
 };
 
-export default function CaeserCipher() {
+export default function ComprehensiveChallenge() {
   const [value, setValue] = React.useState(0);
   const handleChange = (events, newValue) => {
     setValue(newValue);
-    setPath(ChallengeSet1Data[newValue].itembox1);
+    setPath(Challengedata[newValue].itembox1);
   };
 
-  // API POST REQUEST TO THE SERVER
   const { data, error, loading } = useFetchAuth(
     "api/user/environments",
     "POST",
-    { challengeSetSlug: "ciphers", challengeSlug: "letter-to-number" }
+    { challengeSetSlug: "ciphers", challengeSlug: "comprehensive-challenge" }
   );
-
-
+  const [txt, setMarkdown] = useState("");
+  const [path, setPath] = useState("");
+  const [first, setFirst] = useState(0);
 
   //LEARN THIS FOR FUTURE WORK:
   //USESTATE HOOK WILL TRIGGER WHEN THE ASSIGNED FUNCTION IS CALL (EXAMPLE BELOW IS SETMARKDOWN/ SETPATH)
   //USEEFFECT HOOK WILL TRIGGER THE WHEN THE VARIABLE AT THE [] AT BOTTOM IS UPDATE. IN THIS CASE IS "PATH VARIABLE"
-  const [txt, setMarkdown] = useState("");
-  const [path, setPath] = useState("");
-  const [first, setfirst] = useState(0);
+
   useEffect(() => {
     fetch(path)
       .then((res) => res.text())
@@ -76,12 +73,12 @@ export default function CaeserCipher() {
         setMarkdown(text);
       });
   }, [path]);
+  //this is to set first frim rendering for mardown. Without this, it will redner html on first run
   if (loading) return <Spinner />;
   if (error) return <GenericErrorPage />;
-  //this is to set first frim rendering for mardown. Without this, it will redner html on first run
   if (!first) {
-    setPath(ChallengeSet1Data[0].itembox1);
-    setfirst(first + 1);
+    setPath(Challengedata[0].itembox1);
+    setFirst(first + 1);
   }
   return (
     <div style={{ display: "flex", flexDirection: "row", position: "fixed" }}>
@@ -96,8 +93,9 @@ export default function CaeserCipher() {
             scrollButtons="auto"
             aria-label="simple auto tabs example"
           >
-            {ChallengeSet1Data.map((item, index) => {
+            {Challengedata.map((item, index) => {
               return (
+                // <core.Tab key={item.index} label={item.label} {...a11yProps(item.index)} component={Link} to={`${item.topic}/${item.label}`} />
                 <core.Tab
                   key={index}
                   label={item.label}
@@ -108,12 +106,15 @@ export default function CaeserCipher() {
           </core.Tabs>
         </core.AppBar>
 
+        {/* <ReactMarkdown source={Challengedata[value].itembox1}/> */}
+
         <TabPanel
           className="box1"
           value={value}
           index={value}
           style={{ overflowY: "scroll", marginTop: "5px", marginLeft: "5px" }}
         >
+          {/* {Challengedata[value].itembox1}  */}
           <ReactMarkdown
             remarkPlugins={[gfm]}
             children={txt}
