+++ ui/src/pages/challenge-pages/MorseCode.js
@@ -1,20 +1,20 @@
 import React, { useState, useEffect } from "react";
-import "./ChallengeSet1.css";
+import "../../style/ChallengeSet1.css";
 import * as core from "@material-ui/core";
 //FIX THIS DATA FOR EACH FILE
 //CHALLENGE 1 = CAESAR CHALLENGE
 /*----------------------------------------------------------------------------- */
-import { Challengedata } from "./LetterToNumberData";
+import { Challengedata } from "../challenge-data/MorseCodeData";
 /*----------------------------------------------------------------------------- */
 // import {useLocation, Link, useParams} from 'react-router-dom';
 import PropTypes from "prop-types";
-import Terminal from "../components/Terminal";
-import marked from "marked";
+import Terminal from "../../components/Terminal";
+// import marked from "marked";
 import gfm from "remark-gfm";
 import ReactMarkdown from "react-markdown";
-import useFetchAuth from "../useFetchAuth";
-import Spinner from "../components/Spinner";
-import GenericErrorPage from "./error-pages/genericErrorPage";
+import useFetchAuth from "../../useFetchAuth";
+import Spinner from "../../components/Spinner";
+import GenericErrorPage from "../error-pages/genericErrorPage";
 
 function a11yProps(index) {
   return {
@@ -47,23 +47,25 @@ TabPanel.propTypes = {
   value: PropTypes.any.isRequired,
 };
 
-export default function LetterToNumber() {
+export default function MorseCode() {
   const [value, setValue] = React.useState(0);
   const handleChange = (events, newValue) => {
     setValue(newValue);
     setPath(Challengedata[newValue].itembox1);
   };
-  const { data, error, loading } = useFetchAuth(
-    "api/user/environments",
-    "POST",
-    { challengeSetSlug: "ciphers", challengeSlug: "letter-to-number" }
-  );
+
   //LEARN THIS FOR FUTURE WORK:
   //USESTATE HOOK WILL TRIGGER WHEN THE ASSIGNED FUNCTION IS CALL (EXAMPLE BELOW IS SETMARKDOWN/ SETPATH)
   //USEEFFECT HOOK WILL TRIGGER THE WHEN THE VARIABLE AT THE [] AT BOTTOM IS UPDATE. IN THIS CASE IS "PATH VARIABLE"
   const [txt, setMarkdown] = useState("");
   const [path, setPath] = useState("");
-  const [first, setFirst] = useState(0);
+  const [first, setfirst] = useState(0);
+  const { data, error, loading } = useFetchAuth(
+    "api/user/environments",
+    "POST",
+    { challengeSetSlug: "ciphers", challengeSlug: "letter-to-number" }
+  );
+
   useEffect(() => {
     fetch(path)
       .then((res) => res.text())
@@ -71,13 +73,12 @@ export default function LetterToNumber() {
         setMarkdown(text);
       });
   }, [path]);
-
-  //this is to set first frim rendering for mardown. Without this, it will redner html on first run
+  //this is to set first from rendering for markdown. Without this, it will render html on first run
   if (loading) return <Spinner />;
   if (error) return <GenericErrorPage />;
   if (!first) {
     setPath(Challengedata[0].itembox1);
-    setFirst(first + 1);
+    setfirst(first + 1);
   }
   return (
     <div style={{ display: "flex", flexDirection: "row", position: "fixed" }}>
@@ -94,6 +95,7 @@ export default function LetterToNumber() {
           >
             {Challengedata.map((item, index) => {
               return (
+                // <core.Tab key={item.index} label={item.label} {...a11yProps(item.index)} component={Link} to={`${item.topic}/${item.label}`} />
                 <core.Tab
                   key={index}
                   label={item.label}
@@ -104,12 +106,15 @@ export default function LetterToNumber() {
           </core.Tabs>
         </core.AppBar>
 
+        {/* <ReactMarkdown source={Challengedata[value].itembox1}/> */}
+
         <TabPanel
           className="box1"
           value={value}
           index={value}
           style={{ overflowY: "scroll", marginTop: "5px", marginLeft: "5px" }}
         >
+          {/* {Challengedata[value].itembox1}  */}
           <ReactMarkdown
             remarkPlugins={[gfm]}
             children={txt}
