+++ ui/src/pages/challenge-pages/ReverseCipher.js
@@ -1,20 +1,17 @@
 import React, { useState, useEffect } from "react";
-import "./ChallengeSet1.css";
+import "../../style/ChallengeSet1.css";
 import * as core from "@material-ui/core";
-//FIX THIS DATA FOR EACH FILE
-//CHALLENGE 1 = CAESAR CHALLENGE
-/*----------------------------------------------------------------------------- */
-import { Challengedata } from "./MorseCodeData";
-/*----------------------------------------------------------------------------- */
-// import {useLocation, Link, useParams} from 'react-router-dom';
+
+import { Challengedata } from "../challenge-data/ReverseCipherData";
+
 import PropTypes from "prop-types";
-import Terminal from "../components/Terminal";
-// import marked from "marked";
+import Terminal from "../../components/Terminal";
+
 import gfm from "remark-gfm";
 import ReactMarkdown from "react-markdown";
-import useFetchAuth from "../useFetchAuth";
-import Spinner from "../components/Spinner";
-import GenericErrorPage from "./error-pages/genericErrorPage";
+import useFetchAuth from "../../useFetchAuth";
+import Spinner from "../../components/Spinner";
+import GenericErrorPage from "../error-pages/genericErrorPage";
 
 function a11yProps(index) {
   return {
@@ -47,25 +44,23 @@ TabPanel.propTypes = {
   value: PropTypes.any.isRequired,
 };
 
-export default function MorseCode() {
+export default function ReverseCipher() {
   const [value, setValue] = React.useState(0);
   const handleChange = (events, newValue) => {
     setValue(newValue);
     setPath(Challengedata[newValue].itembox1);
   };
-
+  const { data, error, loading } = useFetchAuth(
+    "api/user/environments",
+    "POST",
+    { challengeSetSlug: "ciphers", challengeSlug: "letter-to-number" }
+  );
   //LEARN THIS FOR FUTURE WORK:
   //USESTATE HOOK WILL TRIGGER WHEN THE ASSIGNED FUNCTION IS CALL (EXAMPLE BELOW IS SETMARKDOWN/ SETPATH)
   //USEEFFECT HOOK WILL TRIGGER THE WHEN THE VARIABLE AT THE [] AT BOTTOM IS UPDATE. IN THIS CASE IS "PATH VARIABLE"
   const [txt, setMarkdown] = useState("");
   const [path, setPath] = useState("");
   const [first, setfirst] = useState(0);
-  const { data, error, loading } = useFetchAuth(
-    "api/user/environments",
-    "POST",
-    { challengeSetSlug: "ciphers", challengeSlug: "letter-to-number" }
-  );
-
   useEffect(() => {
     fetch(path)
       .then((res) => res.text())
@@ -73,9 +68,10 @@ export default function MorseCode() {
         setMarkdown(text);
       });
   }, [path]);
-  //this is to set first from rendering for markdown. Without this, it will render html on first run
+
   if (loading) return <Spinner />;
   if (error) return <GenericErrorPage />;
+  //this is to set first frim rendering for mardown. Without this, it will redner html on first run
   if (!first) {
     setPath(Challengedata[0].itembox1);
     setfirst(first + 1);
