+++ ui/src/pages/challenge-pages/WebStructure.js
@@ -1,14 +1,14 @@
 import React,{useState, useEffect} from 'react';
-import './ChallengeSet1.css';
+import '../../style/ChallengeSet1.css';
 import * as core from '@material-ui/core';
 //FIX THIS DATA FOR EACH FILE
 //CHALLENGE 1 = CAESAR CHALLENGE
 /*----------------------------------------------------------------------------- */
-import {Challengedata} from './WebStructureData';
+import {Challengedata} from '../challenge-data/WebStructureData';
 /*----------------------------------------------------------------------------- */
 // import {useLocation, Link, useParams} from 'react-router-dom';
 import PropTypes from 'prop-types';
-import Terminal from '../components/Terminal';
+import Terminal from '../../components/Terminal';
 import marked from 'marked';
 import gfm from 'remark-gfm';
 import ReactMarkdown from 'react-markdown';
