+++ ui/src/style/LeaderBoard.css
